+++ packages/state-manager/src/sagas/messages/sendDeletionMessage/sendDeletionMessage.saga.ts
@@ -4,17 +4,15 @@ import { communitiesSelectors } from '../../communities/communities.selectors'
 import { publicChannelsSelectors } from '../../publicChannels/publicChannels.selectors'
 import { messagesActions } from '../messages.slice'
 import { MessageType, type WriteMessagePayload } from '@quiet/types'
-import { userProfileSelectors } from '../../users/userProfile/userProfile.selectors'
 import { deleteChannelMessage } from '@quiet/common'
 
 export function* sendDeletionMessageSaga(
   action: PayloadAction<ReturnType<typeof messagesActions.sendDeletionMessage>['payload']>
 ): Generator {
-  const { channelId } = action.payload
+  const { channelId, isPublic } = action.payload
   const generalChannel = yield* select(publicChannelsSelectors.generalChannel)
   if (!generalChannel) return
-
-  const user = yield* select(userProfileSelectors.myUserProfile)
+  if (!isPublic) return
 
   const isGeneral = channelId === generalChannel.id