+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Adds one-on-one and group DMs to desktop [#2571](https://github.com/TryQuiet/quiet/issues/2571)
 
+### Fixes
+
+* Don't send deletion message for private channels [#3273](https://github.com/TryQuiet/quiet/issues/3273)
+
 ## [7.3.0]
 
 ### Features